% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expect_neet.R
\name{expect_neet}
\alias{expect_neet}
\title{Test for non-empty thing of expected type}
\usage{
expect_neet(thing_to_test)
}
\description{
A \emph{neet test} tests for \strong{non-empty} thing of \strong{expected type}. This
is what is referred to as a \emph{boundary condition} test in RStudio's primers
(todo: citation).
}
\details{
These expectations are intended to integrate into minimal testing workflow
\emph{for development} of data analyses. When developing a function, we will change parameters,
and structure of the pipeline. These tests
enable the developer to feel reassured the pipeline's functions are
outputting non-empty thing of expected type, while the developer decides on
the best structure for an analysis pipeline.

A \code{character} string will be checked for being of string-length > 1.

A \code{numeric} is checked for not being \code{NA}, \code{NULL}, \code{Inf}, or \code{-Inf}.

A `list`` is checked for being of length > 1.

A \code{data.frame} is checked for having at least one row.

A \code{ggplot} object is checked for successfully running.
}
