% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{workflow}
\alias{workflow}
\title{Worfklow}
\usage{
workflow(progress = 0)
}
\description{
Draw an image of a testing workflow and where the user is up to in the
process.
}
\details{
This was adapted from code snippets posted in this
\href{https://twitter.com/CSchmert/status/1224333501840928770?s=20}{thread}.
}
